perf(cli): split ops maintenance into lazily-dispatched per-command modules
Problem: `ops maintenance` (any leaf, including bare --help) cost
1.6-1.9s cold -- polylogue-20d.2's own PR (#2874) landed the rest of
the interactive-performance slice but explicitly deferred this one,
naming it as a real, tractable gap rather than closing the epic
dishonestly (polylogue-sod7). The single 2954-line
polylogue/cli/commands/maintenance.py module was wrapped in one
_LazyGroup, so `_LazyGroup._resolve()` imported the WHOLE module --
all ~30 top-level imports across all 26 subcommands (ArchiveStore,
blob_gc, blob_integrity, embeddings.reconcile, migration_runner,
archive_init/archive_plan/bootstrap/user_write) -- just to render one
subcommand's --help text.

What changed: converted the module into a package
(polylogue/cli/commands/maintenance/), one submodule per command (or
tightly cohesive command family), following the exact _LazyCommand
pattern already used for the root CLI's own dispatch
(cli/click_command_registration.py) -- __init__.py builds the group
and lazily attaches each subcommand by (module, attr, short_help)
instead of decorating them all inline. Within each submodule, every
heavy runtime-only import (ArchiveStore, blob_gc/blob_integrity/
embeddings.reconcile/migration_runner functions, archive_init/
archive_plan/user_write functions) moved from module scope into the
owning function's body -- safe because `from __future__ import
annotations` already makes this file's type hints lazy, so nothing
needed those imports just to build the Click command object.

Two commands needed a real value (not just a type) at Click-decoration
time for their `--help` output: assertion-export's --kind choices
(AssertionKind) now import directly from polylogue.core.enums, the
same lightweight module archive_tiers.user_write itself sources it
from, instead of paying user_write's own heavier import surface.
preview's --scope choices and embedding-orphan-reconcile's
--quiet-window-seconds default are small literal constants now
hardcoded in the CLI module (mirroring
polylogue.maintenance.preview.ALL_SCOPES /
polylogue.storage.embeddings.reconcile.DEFAULT_QUIET_WINDOW_MS), each
with a dedicated drift-guard test asserting equality against the real
source constant.

migrate-tier is the one command left over budget (~1.04s), documented
in its own module docstring: its `tier` argument's click.Choice(...)
needs DURABLE_MIGRATION_TIERS (derived from ArchiveTier) at decoration
time, and importing polylogue.storage.sqlite.archive_tiers.types --
via that package's own __init__.py, which eagerly imports every
tier's DDL module -- costs ~950ms-1.2s regardless of which specific
archive_tiers name is needed. That is a separate, deeper architectural
gap (the archive_tiers package-init import-weight itself), out of
this refactor's scope; devtools/help_latency_probe.py keeps it
informational with that reasoning inline, not silently promoted.

Verification:
- devtools bench help-latency: all 13 required targets green (root
  through ops-maintenance-archive-read, both AC-named targets),
  ops-maintenance-migrate-tier informational at 1043.7ms/700ms.
- python -X importtime -m polylogue.cli ops maintenance archive-read
  --help: zero occurrences of ArchiveStore/blob_gc/blob_integrity/
  embeddings.reconcile/migration_runner/surfaces.payloads in the trace
  (196 real import entries).
- devtools test tests/unit/cli/test_archive_maintenance_cli.py
  tests/unit/cli/test_maintenance_registration.py
  tests/unit/cli/test_paths.py tests/unit/maintenance/
  tests/benchmarks/test_archive_maintenance.py: 251 passed, 5 failed.
  All 5 failures independently reproduced identically against
  unmodified origin/master in a clean worktree (sqlite trigger
  already-exists + embedding-reconcile index-generation-pointer +
  one assertion-ordering mismatch) -- confirmed pre-existing and
  unrelated, not caused by this change.
- devtools verify --quick: exit 0, all 16 static/generated gates
  green.
- mypy --strict polylogue/cli/commands/maintenance/
  devtools/help_latency_probe.py: no issues, 18 source files.
- devtools render topology-projection / topology-status: regenerated
  for the new package (required per CLAUDE.md when adding modules
  under polylogue/).

Ref polylogue-sod7.
