+++ docs/topology-status.md
@@ -28,12 +28,12 @@ Generated by `devtools render topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 798
+- **Stable** (no move scoped): 814
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 18
 - **TBD** (cell needs explicit assignment): 6
-- **Total declared**: 952
-- **Realized polylogue/**/*.py**: 952 files declared
+- **Total declared**: 968
+- **Realized polylogue/**/*.py**: 968 files declared
 
 ### TBD cells (require explicit routing)
