+++ polylogue/cli/commands/maintenance/_plan.py
@@ -1,0 +1,99 @@
+"""``maintenance plan``: dry-run summary of what a backfill would touch."""
+
+from __future__ import annotations
+
+import json
+
+import click
+
+from polylogue.cli.commands.maintenance._shared import _apply_scope_filter_options, _build_scope_filter
+from polylogue.cli.shared.types import AppEnv
+from polylogue.config import Config
+from polylogue.logging import configure_logging
+from polylogue.maintenance.envelope import envelope_from_operation
+from polylogue.maintenance.planner import preview_backfill
+from polylogue.maintenance.targets import MAINTENANCE_TARGET_NAMES, build_maintenance_target_catalog
+from polylogue.paths import archive_root, render_root
+
+_MAINTENANCE_TARGET_HELP = build_maintenance_target_catalog().help_text()
+
+
+@click.command("plan")
+@click.option(
+    "--target",
+    "targets",
+    multiple=True,
+    type=click.Choice(MAINTENANCE_TARGET_NAMES),
+    help=_MAINTENANCE_TARGET_HELP,
+)
+@click.option(
+    "--output-format",
+    "output_format",
+    type=click.Choice(["plain", "json"]),
+    default="plain",
+    show_default=True,
+    help="Output format. ``json`` emits the shared MaintenanceOperationEnvelope.",
+)
+@_apply_scope_filter_options
+@click.pass_obj
+def plan_command(
+    env: AppEnv,
+    targets: tuple[str, ...],
+    output_format: str,
+    session_ids: tuple[str, ...],
+    origin: str | None,
+    source_family: str | None,
+    source_root: str | None,
+    since: str | None,
+    until: str | None,
+    failure_kind: str | None,
+    parser_version: str | None,
+) -> None:
+    """Dry-run summary: show what would be rebuilt without executing.
+
+    Displays affected rows and estimated time for each target.
+    Read-only — no mutations are performed.
+    """
+    configure_logging()
+    config = Config(
+        archive_root=archive_root(),
+        render_root=render_root(),
+        sources=[],  # maintenance doesn't need source acquisition
+    )
+    scope_filter = _build_scope_filter(
+        session_ids=session_ids,
+        origin=origin,
+        source_family=source_family,
+        source_root=source_root,
+        since=since,
+        until=until,
+        failure_kind=failure_kind,
+        parser_version=parser_version,
+    )
+    result = preview_backfill(config, targets=targets, scope_filter=scope_filter)
+
+    if output_format == "json":
+        envelope = envelope_from_operation(result, origin="cli", mode="preview")
+        click.echo(json.dumps(envelope.to_dict(), indent=2, sort_keys=True))
+        return
+
+    click.echo(f"Operation: {result.operation_id}")
+    click.echo(f"Targets:  {', '.join(result.targets) if result.targets else 'all'}")
+    click.echo(f"Affected: {result.affected_rows:,} rows")
+    if result.estimated_time_s > 0:
+        click.echo(f"Estimate: ~{result.estimated_time_s:.1f}s")
+
+    if result.results:
+        click.echo("\nPer-target preview:")
+        for r in result.results:
+            name = r.get("name", "unknown")
+            issue_count = r.get("issue_count", 0)
+            healthy = r.get("healthy", True)
+            detail = r.get("detail", "")
+            status_str = "OK" if healthy else f"{issue_count:,} issues"
+            click.echo(f"  {name}: {status_str}")
+            if detail and not healthy:
+                click.echo(f"    {detail}")
+
+    if result.error:
+        click.echo(f"\nError: {result.error}", err=True)
