+++ tests/unit/cli/test_archive_maintenance_cli.py
@@ -10,7 +10,7 @@ import pytest
 from click.testing import CliRunner
 
 from polylogue.cli.click_app import cli
-from polylogue.cli.commands import maintenance
+from polylogue.cli.commands.maintenance import _rebuild_index as maintenance_rebuild_index
 from polylogue.config import Config
 from polylogue.core.enums import Provider
 from polylogue.maintenance.replay import rebuild_index_from_source
@@ -841,6 +841,20 @@ def _seed_orphan_embedding_row(archive_root: Path) -> tuple[str, str]:
     return session_id, orphan_message_id
 
 
+def test_embedding_orphan_reconcile_default_quiet_window_matches_reconcile_module() -> None:
+    """The CLI's hardcoded --help default (polylogue-sod7) must not drift from the real constant.
+
+    _embeddings.py hardcodes _DEFAULT_QUIET_WINDOW_SECONDS instead of importing
+    DEFAULT_QUIET_WINDOW_MS from polylogue.storage.embeddings.reconcile, so
+    that constant -- and its heavy import chain -- isn't paid on the
+    `--help` path. This test is the drift guard for that duplication.
+    """
+    from polylogue.cli.commands.maintenance._embeddings import _DEFAULT_QUIET_WINDOW_SECONDS
+    from polylogue.storage.embeddings.reconcile import DEFAULT_QUIET_WINDOW_MS
+
+    assert _DEFAULT_QUIET_WINDOW_SECONDS == DEFAULT_QUIET_WINDOW_MS // 1000
+
+
 def test_embedding_orphan_reconcile_cli_dry_run_keeps_rows(
     cli_workspace: dict[str, Path],
     cli_runner: CliRunner,
@@ -1068,7 +1082,10 @@ def test_archive_init_cli_executes_confirmed_initialization(
             ),
         )
 
-    monkeypatch.setattr(maintenance, "initialize_archive_tier_files_from_plan", fake_init)
+    monkeypatch.setattr(
+        "polylogue.storage.sqlite.archive_tiers.archive_init.initialize_archive_tier_files_from_plan",
+        fake_init,
+    )
 
     result = cli_runner.invoke(
         cli,
@@ -1679,7 +1696,7 @@ def test_archive_read_cli_lists_archive_sessions(
             ]
 
     monkeypatch.setattr(
-        "polylogue.cli.commands.maintenance.ArchiveStore.open_existing",
+        "polylogue.storage.sqlite.archive_tiers.archive.ArchiveStore.open_existing",
         classmethod(lambda cls, root: FakeArchiveStore()),
     )
 
@@ -1752,7 +1769,7 @@ def test_archive_read_cli_searches_archive_blocks(
             ]
 
     monkeypatch.setattr(
-        "polylogue.cli.commands.maintenance.ArchiveStore.open_existing",
+        "polylogue.storage.sqlite.archive_tiers.archive.ArchiveStore.open_existing",
         classmethod(lambda cls, root: FakeArchiveStore()),
     )
 
@@ -1795,13 +1812,13 @@ def test_rebuild_index_source_replay_expands_every_execution_selection_to_author
     monkeypatch: pytest.MonkeyPatch,
     selection_args: list[str],
 ) -> None:
-    monkeypatch.setattr("polylogue.cli.commands.maintenance._count_source_raw_sessions", lambda _root: 4)
+    monkeypatch.setattr("polylogue.cli.commands.maintenance._rebuild_index._count_source_raw_sessions", lambda _root: 4)
     monkeypatch.setattr(
-        "polylogue.cli.commands.maintenance._all_index_rebuild_raw_ids",
+        "polylogue.cli.commands.maintenance._rebuild_index._all_index_rebuild_raw_ids",
         lambda _root: ["raw-parent", "raw-child"],
     )
     monkeypatch.setattr(
-        "polylogue.cli.commands.maintenance._missing_index_raw_ids",
+        "polylogue.cli.commands.maintenance._rebuild_index._missing_index_raw_ids",
         lambda _root: ["raw-parent", "raw-child"],
     )
 
@@ -1875,7 +1892,7 @@ def test_all_index_rebuild_raw_ids_uses_source_acquisition_order(
                 (raw_id, raw_id, f"/tmp/{raw_id}.jsonl", acquired_at_ms),
             )
 
-    assert maintenance._all_index_rebuild_raw_ids(cli_workspace["archive_root"]) == [
+    assert maintenance_rebuild_index._all_index_rebuild_raw_ids(cli_workspace["archive_root"]) == [
         "raw-parent",
         "raw-sibling-a",
         "raw-sibling-b",
@@ -1917,7 +1934,9 @@ def test_rebuild_index_explicit_raw_ids_remain_inspectable_in_plan_mode(
     cli_runner: CliRunner,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    monkeypatch.setattr("polylogue.cli.commands.maintenance._count_source_raw_sessions", lambda _root: 10)
+    monkeypatch.setattr(
+        "polylogue.cli.commands.maintenance._rebuild_index._count_source_raw_sessions", lambda _root: 10
+    )
 
     result = cli_runner.invoke(
         cli,
@@ -1973,9 +1992,9 @@ def test_rebuild_index_filters_selected_rows_by_blob_size(
                 (raw_id, raw_id, f"/tmp/{raw_id}.jsonl", blob_size, acquired_at_ms),
             )
 
-    monkeypatch.setattr("polylogue.cli.commands.maintenance._count_source_raw_sessions", lambda _root: 2)
+    monkeypatch.setattr("polylogue.cli.commands.maintenance._rebuild_index._count_source_raw_sessions", lambda _root: 2)
     monkeypatch.setattr(
-        "polylogue.cli.commands.maintenance._missing_index_raw_ids",
+        "polylogue.cli.commands.maintenance._rebuild_index._missing_index_raw_ids",
         lambda _root: ["raw-large", "raw-small"],
     )
     result = cli_runner.invoke(
