+++ tests/unit/maintenance/test_scope_filter.py
@@ -291,9 +291,9 @@ def _capture_cli_preview(operation: BackfillOperation, tmp_path: Path) -> dict[s
     config_obj = Config(archive_root=archive, render_root=render, sources=[])
 
     with (
-        patch("polylogue.cli.commands.maintenance.preview_backfill", return_value=operation),
-        patch("polylogue.cli.commands.maintenance.archive_root", return_value=archive),
-        patch("polylogue.cli.commands.maintenance.render_root", return_value=render),
+        patch("polylogue.cli.commands.maintenance._plan.preview_backfill", return_value=operation),
+        patch("polylogue.cli.commands.maintenance._plan.archive_root", return_value=archive),
+        patch("polylogue.cli.commands.maintenance._plan.render_root", return_value=render),
     ):
         result = runner.invoke(
             maintenance_group,
